pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "logsynth"
version = "0.1.0"
description = "Flexible synthetic log generator with YAML templates, rate control, and LLM-powered template synthesis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Lance" }
]
keywords = ["logging", "synthetic", "generator", "testing", "devops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
logsynth = "logsynth.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["logsynth"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
